refactor: 이미지 파일 형식 변경
